+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/MainActivity.kt
@@ -144,12 +144,6 @@ class MainActivity : AppCompatActivity() {
             setSupportActionBar(this)
         }
 
-        val apiType = if (BuildConfig.ENABLE_SBS) {
-            ApiType.SBS
-        } else {
-            ApiType.GEOCODING
-        }
-
         searchResultsView = findViewById<SearchResultsView>(R.id.search_results_view).apply {
             initialize(
                 SearchResultsView.Configuration(CommonSearchViewConfiguration(DistanceUnitType.IMPERIAL))
@@ -158,7 +152,7 @@ class MainActivity : AppCompatActivity() {
         }
 
         val searchEngine = SearchEngine.createSearchEngineWithBuiltInDataProviders(
-            apiType = apiType,
+            apiType = ApiType.SearchBox,
             settings = SearchEngineSettings(getString(R.string.mapbox_access_token))
         )