+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/api/JapanSearchKotlinExampleActivity.kt
@@ -57,7 +57,7 @@ class JapanSearchKotlinExampleActivity : Activity() {
         super.onCreate(savedInstanceState)
 
         searchEngine = SearchEngine.createSearchEngineWithBuiltInDataProviders(
-            ApiType.SBS,
+            ApiType.SearchBox,
             SearchEngineSettings(getString(R.string.mapbox_access_token))
         )