+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/api/ReverseGeocodingKotlinExampleActivity.kt
@@ -38,7 +38,7 @@ class ReverseGeocodingKotlinExampleActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
 
         searchEngine = SearchEngine.createSearchEngineWithBuiltInDataProviders(
-            ApiType.SBS,
+            ApiType.SearchBox,
             SearchEngineSettings(getString(R.string.mapbox_access_token))
         )