+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/CategorySearchIntegrationTest.kt
@@ -84,7 +84,7 @@ internal class CategorySearchIntegrationTest : BaseTest() {
         )
 
         searchEngine = createSearchEngineWithBuiltInDataProvidersBlocking(
-            apiType = ApiType.SBS,
+            apiType = ApiType.SearchBox,
             settings = searchEngineSettings,
         )