+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/SearchEngineIntegrationTest.kt
@@ -107,7 +107,7 @@ internal class SearchEngineIntegrationTest : BaseTest() {
         )
 
         searchEngine = createSearchEngineWithBuiltInDataProvidersBlocking(
-            apiType = ApiType.SBS,
+            apiType = ApiType.SearchBox,
             settings = searchEngineSettings,
         )
 
@@ -910,7 +910,7 @@ internal class SearchEngineIntegrationTest : BaseTest() {
 
     @Test
     fun testGeocodingCzechAddressFormatting() {
-        searchEngine = SearchEngine.createSearchEngine(ApiType.GEOCODING, searchEngineSettings)
+        searchEngine = SearchEngine.createSearchEngine(ApiType.Geocoding, searchEngineSettings)
 
         mockServer.enqueue(createSuccessfulResponse("geocoding_responses/suggestions-successful-czech.json"))
 
@@ -968,7 +968,7 @@ internal class SearchEngineIntegrationTest : BaseTest() {
 
     @Test
     fun testMetadataForGeocodingAPI() {
-        searchEngine = SearchEngine.createSearchEngine(ApiType.GEOCODING, searchEngineSettings)
+        searchEngine = SearchEngine.createSearchEngine(ApiType.Geocoding, searchEngineSettings)
 
         mockServer.enqueue(createSuccessfulResponse("geocoding_responses/suggestions.json"))