+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/ApiType.kt
@@ -1,30 +1,30 @@
 package com.mapbox.search
 
-import com.mapbox.search.Reserved.Flags
 import com.mapbox.search.base.core.CoreApiType
 
 /**
  * The type of the API used by one of the Search Engines.
  */
-public enum class ApiType {
+public abstract class ApiType {
 
     /**
      * Geocoding API.
+     * @see [Geocoding Api documentation](https://docs.mapbox.com/api/search/geocoding/)
      */
-    GEOCODING,
+    public object Geocoding : ApiType()
 
     /**
-     * Single Box Search API.
+     * Search Box API.
+     * @see [Search Box Api documentation](https://docs.mapbox.com/api/search/search-box/)
      */
-    @Reserved(Flags.SBS)
-    // TODO FIXME rename
-    SBS,
+    public object SearchBox : ApiType()
 }
 
 @JvmSynthetic
 internal fun ApiType.mapToCore(): CoreApiType {
     return when (this) {
-        ApiType.GEOCODING -> CoreApiType.GEOCODING
-        ApiType.SBS -> CoreApiType.SEARCH_BOX
+        is ApiType.Geocoding -> CoreApiType.GEOCODING
+        is ApiType.SearchBox -> CoreApiType.SEARCH_BOX
+        else -> error("Unknown Api Type")
     }
 }